test(doc): isolate config reload fixture
