+++ packages/rstack/tests/config/reload-doc-config/docs/index.md
@@ -1,0 +1,1 @@
+# Reload doc config
